src/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
Deprecated==1.2.13
Django==4.1.3
django-cors-headers==3.13.0
django-extensions==3.2.1
djangorestframework==3.14.0
dnspython==2.2.1
drf-yasg==1.21.4
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.4.2
MarkupSafe==2.1.1
mongoengine==0.24.2
oauthlib==3.2.2
packaging==21.3
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
pymongo==4.3.2
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
